pyproject.toml
[build-system]
requires = ["setuptools>=69,<75", "wheel>=0.43,<1"]
build-backend = "setuptools.build_meta"

[project]
name = "backtestforecast"
version = "0.1.0"
description = "BacktestForecast.com monorepo Python application layer."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "alembic>=1.13,<2.0",
  "psycopg[binary]>=3.2,<4.0",
  "pydantic>=2.8,<3.0",
  "pydantic-settings>=2.4,<3.0",
  "celery[redis]>=5.4,<6.0",
  "redis>=5.0,<6.0",
  "httpx>=0.27,<1.0",
  "PyJWT[crypto]>=2.10,<3.0",
  "structlog>=24.4,<26.0",
  "stripe>=11.0,<12.0",
  "reportlab>=4.2,<5.0",
  "prometheus-client>=0.21,<1.0",
  "sse-starlette>=2.0,<4.0",
  "celery-redbeat>=2.2,<3.0"
]

[project.optional-dependencies]
s3 = [
  "boto3>=1.34,<2.0"
]
sentry = [
  "sentry-sdk[fastapi,celery]>=2.0,<3.0"
]
otel = [
  "opentelemetry-api>=1.20,<2.0",
  "opentelemetry-sdk>=1.20,<2.0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.20,<2.0",
  "opentelemetry-instrumentation-fastapi>=0.44b0",
  "opentelemetry-instrumentation-sqlalchemy>=0.44b0",
  "opentelemetry-instrumentation-redis>=0.44b0",
  "opentelemetry-instrumentation-httpx>=0.44b0",
  "opentelemetry-instrumentation-celery>=0.44b0",
]
dev = [
  "pytest>=8.3,<10.0",
  "pytest-asyncio>=0.24,<2.0",
  "pytest-cov>=6.0,<7.0",
  "time-machine>=2.16,<3.0",
  "respx>=0.22,<1.0",
  "ruff>=0.7,<1.0",
  "mypy>=1.13,<2.0",
  "locust>=2.29,<3.0",
  "pytest-timeout>=2.3,<3.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "T20",  # flake8-print (catch stray print() calls)
    "RUF",  # Ruff-specific rules
    "S",    # flake8-bandit security checks
]
ignore = [
    "E501",   # line length handled by formatter
    "B008",   # function call in default argument (common in FastAPI Depends())
    "RUF012", # mutable class variable annotation (conflicts with SQLAlchemy)
]

[tool.ruff.lint.per-file-ignores]
"alembic/versions/*" = ["E501", "UP", "B007", "RUF002", "T201"]
"tests/**" = ["T20", "E402", "E741", "S101", "S105", "S106", "S107", "S108", "S110", "S112", "S311", "S603", "S607", "S608", "RUF001", "RUF002", "RUF003", "B017", "SIM102", "SIM115", "SIM117"]
"scripts/**" = ["T20", "E402", "S101", "S105", "S106", "S603", "S607", "SIM102"]
"src/backtestforecast/billing/entitlements.py" = ["S110"]
"src/backtestforecast/config.py" = ["S105", "S106", "S110"]
"src/backtestforecast/exports/__init__.py" = ["E402", "S311"]
"src/backtestforecast/exports/storage.py" = ["E402", "S311"]
"src/backtestforecast/feature_flags.py" = ["S110"]
"src/backtestforecast/integrations/massive_client.py" = ["S311"]
"src/backtestforecast/management/backfill_metrics.py" = ["S110"]
"src/backtestforecast/observability/metrics.py" = ["E402"]
"src/backtestforecast/pipeline/deep_analysis.py" = ["E402", "S110"]
"src/backtestforecast/pipeline/service.py" = ["S110"]
"src/backtestforecast/schemas/json_shapes.py" = ["S110"]
"src/backtestforecast/security/rate_limits.py" = ["S311"]
"src/backtestforecast/services/scans.py" = ["E402", "S311"]
"src/backtestforecast/services/sweeps.py" = ["E402"]
"src/backtestforecast/sweeps/constraints.py" = ["S311"]
"src/backtestforecast/sweeps/genetic.py" = ["S311"]

[tool.ruff.lint.isort]
known-first-party = ["backtestforecast"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers"
markers = [
    "smoke: full-stack smoke tests (run with pytest -m smoke)",
    "integration: tests that require a running database",
    "e2e: end-to-end tests",
    "real_worker: tests that require a live Redis broker and a real Celery worker",
    "slow: long-running tests",
    "asyncio: async tests (registered by pytest-asyncio)",
    "target_assertion: critical regression tests that must explicitly mark when they reach the business assertion they are intended to protect",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:celery.*",
    "ignore::DeprecationWarning:kombu.*",
]
timeout = 120
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/alembic/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true

# Per-module coverage enforcement.  CI runs with --cov-fail-under=80 globally;
# these thresholds catch coverage regressions in critical modules that the
# global average might mask.
# To check locally: pytest --cov=src --cov-report=term-missing
# Module-specific checks are enforced by the test below.

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
warn_return_any = true
warn_unreachable = true
disallow_any_generics = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "kombu.*",
    "redbeat.*",
    "reportlab.*",
    "stripe.*",
    "boto3.*",
    "botocore.*",
]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
